+++ apps/vs-agent/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './ServerConfig'
 export * from './logger'
 export * from './setupAgent'
-export * from './webhookEvent'
